discord-bot/requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Utilities
structlog>=24.0.0
